// assets/Scripts/enum/Enum.ts
export enum GAME_EVENT_ENUM {
    PLAY_AUDIO = "PLAY_AUDIO",
    PLAY_DDZ_GAME = "PLAY_DDZ_GAME",
}

export enum DDZ_EFFECT_ENUM {
    PLAYER_READY = "player_ready",
    PLAYER_BID = "player_bid",
    PLAYER_PLAY = "player_play",
    PLAYER_PASS = "player_pass",
    GAME_STATE_CHANGED = "game_state_changed",
    GAME_OVER = "game_over",
}

export enum AUDIO_EFFECT_ENUM {
    CLICK_BUTTON,
    CLEAR,
    BUYAO,
    CHUPAI,
    JIAODIZHU,
    BUJIAO,
    FAPAI,
}

export enum GAME_SCENE_ENUM {
    MENU = "Menu",
    GAME = "Game"
}